perf: isolate kernel types from graph
